% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{BaseRenderer}
\alias{BaseRenderer}
\title{BaseRenderer}
\description{
...
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseRenderer-new}{\code{BaseRenderer$new()}}
\item \href{#method-BaseRenderer-outDir}{\code{BaseRenderer$outDir()}}
\item \href{#method-BaseRenderer-colorHtml}{\code{BaseRenderer$colorHtml()}}
\item \href{#method-BaseRenderer-emojiHtml}{\code{BaseRenderer$emojiHtml()}}
\item \href{#method-BaseRenderer-specialBlockHtml}{\code{BaseRenderer$specialBlockHtml()}}
\item \href{#method-BaseRenderer-extRefHtml}{\code{BaseRenderer$extRefHtml()}}
\item \href{#method-BaseRenderer-intRefHtml}{\code{BaseRenderer$intRefHtml()}}
\item \href{#method-BaseRenderer-id2url}{\code{BaseRenderer$id2url()}}
\item \href{#method-BaseRenderer-makeAll}{\code{BaseRenderer$makeAll()}}
\item \href{#method-BaseRenderer-makeToc}{\code{BaseRenderer$makeToc()}}
\item \href{#method-BaseRenderer-makeMaterials}{\code{BaseRenderer$makeMaterials()}}
\item \href{#method-BaseRenderer-makeLectures}{\code{BaseRenderer$makeLectures()}}
\item \href{#method-BaseRenderer-makeTasks}{\code{BaseRenderer$makeTasks()}}
\item \href{#method-BaseRenderer-docUrl}{\code{BaseRenderer$docUrl()}}
\item \href{#method-BaseRenderer-mapTmpRmdFile}{\code{BaseRenderer$mapTmpRmdFile()}}
\item \href{#method-BaseRenderer-mapOutHtmlFile}{\code{BaseRenderer$mapOutHtmlFile()}}
\item \href{#method-BaseRenderer-mapOutMaterialFile}{\code{BaseRenderer$mapOutMaterialFile()}}
\item \href{#method-BaseRenderer-clearDir}{\code{BaseRenderer$clearDir()}}
\item \href{#method-BaseRenderer-makeDoc}{\code{BaseRenderer$makeDoc()}}
\item \href{#method-BaseRenderer-copyDoc}{\code{BaseRenderer$copyDoc()}}
\item \href{#method-BaseRenderer-renderDoc}{\code{BaseRenderer$renderDoc()}}
\item \href{#method-BaseRenderer-clone}{\code{BaseRenderer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$new(outDir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-outDir"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-outDir}{}}}
\subsection{Method \code{outDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$outDir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-colorHtml"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-colorHtml}{}}}
\subsection{Method \code{colorHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$colorHtml(..., color)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-emojiHtml"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-emojiHtml}{}}}
\subsection{Method \code{emojiHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$emojiHtml(id, text = id, color = "red")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-specialBlockHtml"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-specialBlockHtml}{}}}
\subsection{Method \code{specialBlockHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$specialBlockHtml(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-extRefHtml"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-extRefHtml}{}}}
\subsection{Method \code{extRefHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$extRefHtml(..., url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-intRefHtml"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-intRefHtml}{}}}
\subsection{Method \code{intRefHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$intRefHtml(..., url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-id2url"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-id2url}{}}}
\subsection{Method \code{id2url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$id2url(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-makeAll"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-makeAll}{}}}
\subsection{Method \code{makeAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeAll(course, makeZip = TRUE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-makeToc"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-makeToc}{}}}
\subsection{Method \code{makeToc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeToc(course, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-makeMaterials"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-makeMaterials}{}}}
\subsection{Method \code{makeMaterials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeMaterials(course, materialIds = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-makeLectures"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-makeLectures}{}}}
\subsection{Method \code{makeLectures()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeLectures(course, lectureIds = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-makeTasks"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-makeTasks}{}}}
\subsection{Method \code{makeTasks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeTasks(course, enableCode, lectureIds = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-docUrl"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-docUrl}{}}}
\subsection{Method \code{docUrl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$docUrl(doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-mapTmpRmdFile"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-mapTmpRmdFile}{}}}
\subsection{Method \code{mapTmpRmdFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$mapTmpRmdFile(doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-mapOutHtmlFile"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-mapOutHtmlFile}{}}}
\subsection{Method \code{mapOutHtmlFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$mapOutHtmlFile(doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-mapOutMaterialFile"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-mapOutMaterialFile}{}}}
\subsection{Method \code{mapOutMaterialFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$mapOutMaterialFile(doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-clearDir"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-clearDir}{}}}
\subsection{Method \code{clearDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$clearDir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-makeDoc"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-makeDoc}{}}}
\subsection{Method \code{makeDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeDoc(course, doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-copyDoc"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-copyDoc}{}}}
\subsection{Method \code{copyDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$copyDoc(course, doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-renderDoc"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-renderDoc}{}}}
\subsection{Method \code{renderDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$renderDoc(course, doc, quiet = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseRenderer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseRenderer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
