% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{BrightspaceRenderer}
\alias{BrightspaceRenderer}
\title{BrightspaceRenderer}
\description{
...
}
\section{Super class}{
\code{\link[rooc:BaseRenderer]{rooc::BaseRenderer}} -> \code{BrightspaceRenderer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BrightspaceRenderer-clone}{\code{BrightspaceRenderer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="clearDir"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-clearDir'><code>rooc::BaseRenderer$clearDir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="colorHtml"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-colorHtml'><code>rooc::BaseRenderer$colorHtml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="copyDoc"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-copyDoc'><code>rooc::BaseRenderer$copyDoc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="docUrl"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-docUrl'><code>rooc::BaseRenderer$docUrl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="emojiHtml"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-emojiHtml'><code>rooc::BaseRenderer$emojiHtml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="extRefHtml"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-extRefHtml'><code>rooc::BaseRenderer$extRefHtml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="id2url"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-id2url'><code>rooc::BaseRenderer$id2url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="initialize"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-initialize'><code>rooc::BaseRenderer$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="intRefHtml"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-intRefHtml'><code>rooc::BaseRenderer$intRefHtml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeAll"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-makeAll'><code>rooc::BaseRenderer$makeAll()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeDoc"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-makeDoc'><code>rooc::BaseRenderer$makeDoc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeLectures"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-makeLectures'><code>rooc::BaseRenderer$makeLectures()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeMaterials"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-makeMaterials'><code>rooc::BaseRenderer$makeMaterials()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeTasks"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-makeTasks'><code>rooc::BaseRenderer$makeTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeToc"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-makeToc'><code>rooc::BaseRenderer$makeToc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="mapOutHtmlFile"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-mapOutHtmlFile'><code>rooc::BaseRenderer$mapOutHtmlFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="mapOutMaterialFile"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-mapOutMaterialFile'><code>rooc::BaseRenderer$mapOutMaterialFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="mapTmpRmdFile"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-mapTmpRmdFile'><code>rooc::BaseRenderer$mapTmpRmdFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="outDir"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-outDir'><code>rooc::BaseRenderer$outDir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="renderDoc"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-renderDoc'><code>rooc::BaseRenderer$renderDoc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="specialBlockHtml"><a href='../../rooc/html/BaseRenderer.html#method-BaseRenderer-specialBlockHtml'><code>rooc::BaseRenderer$specialBlockHtml()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BrightspaceRenderer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BrightspaceRenderer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrightspaceRenderer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
