% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/language.R
\name{code}
\alias{code}
\alias{koodi}
\title{Show the code for the solution to the current problem}
\usage{
code()

koodi()
}
\description{
Show the code for the solution to the current problem
}
