% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{CopiedDoc}
\alias{CopiedDoc}
\title{CopiedDoc}
\description{
...
}
\section{Super class}{
\code{\link[rooc:Doc]{rooc::Doc}} -> \code{CopiedDoc}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CopiedDoc$new()}}
\item \href{#method-path}{\code{CopiedDoc$path()}}
\item \href{#method-outPath}{\code{CopiedDoc$outPath()}}
\item \href{#method-url}{\code{CopiedDoc$url()}}
\item \href{#method-clone}{\code{CopiedDoc$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="Doc" data-id="id">}\href{../../rooc/html/Doc.html#method-id}{\code{rooc::Doc$id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="Doc" data-id="label">}\href{../../rooc/html/Doc.html#method-label}{\code{rooc::Doc$label()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopiedDoc$new(path, outPath = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path"></a>}}
\if{latex}{\out{\hypertarget{method-path}{}}}
\subsection{Method \code{path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopiedDoc$path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-outPath"></a>}}
\if{latex}{\out{\hypertarget{method-outPath}{}}}
\subsection{Method \code{outPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopiedDoc$outPath()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-url"></a>}}
\if{latex}{\out{\hypertarget{method-url}{}}}
\subsection{Method \code{url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopiedDoc$url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopiedDoc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
