% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CourseSite.R
\docType{class}
\name{CourseSite}
\alias{CourseSite}
\title{CourseSite}
\format{An \code{\link{R6Class}} generator object}
\description{
Collection of methods to build a web-site with rmarkdown::render_site().
}
\section{Methods}{

\describe{
  \item{...}{...}
}
}

\examples{


}
\seealso{
\code{\link{CourseSite}}
}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CourseSite$new()}}
\item \href{#method-build}{\code{CourseSite$build()}}
\item \href{#method-url}{\code{CourseSite$url()}}
\item \href{#method-path}{\code{CourseSite$path()}}
\item \href{#method-view}{\code{CourseSite$view()}}
\item \href{#method-courseSummary}{\code{CourseSite$courseSummary()}}
\item \href{#method-slots}{\code{CourseSite$slots()}}
\item \href{#method-clone}{\code{CourseSite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CourseSite$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build"></a>}}
\subsection{Method \code{build()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CourseSite$build()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-url"></a>}}
\subsection{Method \code{url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CourseSite$url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path"></a>}}
\subsection{Method \code{path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CourseSite$path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-view"></a>}}
\subsection{Method \code{view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CourseSite$view()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-courseSummary"></a>}}
\subsection{Method \code{courseSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CourseSite$courseSummary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-slots"></a>}}
\subsection{Method \code{slots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CourseSite$slots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CourseSite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
