% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_if_missing.R
\name{install_if_missing}
\alias{install_if_missing}
\title{Install and load missing packages}
\usage{
install_if_missing(package_names = character())
}
\arguments{
\item{package_names}{A vector of package names.}
}
\value{
No return value.
}
\description{
Install and load missing packages
}
