% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue.R
\name{issue}
\alias{issue}
\title{Submit an issue about an R course}
\usage{
issue(id = getOption("rcourse_id", default = select()))
}
\arguments{
\item{id}{The course id.}
}
\value{
No return value.
}
\description{
Submit an issue about an R course
}
