% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{Lecture}
\alias{Lecture}
\title{Lecture}
\description{
...
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Lecture$new()}}
\item \href{#method-id}{\code{Lecture$id()}}
\item \href{#method-label}{\code{Lecture$label()}}
\item \href{#method-rmdFile}{\code{Lecture$rmdFile()}}
\item \href{#method-hasTasks}{\code{Lecture$hasTasks()}}
\item \href{#method-min}{\code{Lecture$min()}}
\item \href{#method-asTibble}{\code{Lecture$asTibble()}}
\item \href{#method-clone}{\code{Lecture$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$new(id, label = NULL, rmdFile = NULL, hasTasks = TRUE, min = NA, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-id"></a>}}
\if{latex}{\out{\hypertarget{method-id}{}}}
\subsection{Method \code{id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label"></a>}}
\if{latex}{\out{\hypertarget{method-label}{}}}
\subsection{Method \code{label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$label()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rmdFile"></a>}}
\if{latex}{\out{\hypertarget{method-rmdFile}{}}}
\subsection{Method \code{rmdFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$rmdFile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasTasks"></a>}}
\if{latex}{\out{\hypertarget{method-hasTasks}{}}}
\subsection{Method \code{hasTasks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$hasTasks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-min"></a>}}
\if{latex}{\out{\hypertarget{method-min}{}}}
\subsection{Method \code{min()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$min()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asTibble"></a>}}
\if{latex}{\out{\hypertarget{method-asTibble}{}}}
\subsection{Method \code{asTibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$asTibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
