% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{Material}
\alias{Material}
\title{Material}
\description{
...
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Material-new}{\code{Material$new()}}
\item \href{#method-Material-id}{\code{Material$id()}}
\item \href{#method-Material-label}{\code{Material$label()}}
\item \href{#method-Material-path}{\code{Material$path()}}
\item \href{#method-Material-outPath}{\code{Material$outPath()}}
\item \href{#method-Material-asTibble}{\code{Material$asTibble()}}
\item \href{#method-Material-clone}{\code{Material$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-new"></a>}}
\if{latex}{\out{\hypertarget{method-Material-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$new(id, label, path, outPath)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-id"></a>}}
\if{latex}{\out{\hypertarget{method-Material-id}{}}}
\subsection{Method \code{id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-label"></a>}}
\if{latex}{\out{\hypertarget{method-Material-label}{}}}
\subsection{Method \code{label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$label()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-path"></a>}}
\if{latex}{\out{\hypertarget{method-Material-path}{}}}
\subsection{Method \code{path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-outPath"></a>}}
\if{latex}{\out{\hypertarget{method-Material-outPath}{}}}
\subsection{Method \code{outPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$outPath()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-asTibble"></a>}}
\if{latex}{\out{\hypertarget{method-Material-asTibble}{}}}
\subsection{Method \code{asTibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$asTibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Material-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
