% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics}
\alias{metrics}
\title{Stock metrics}
\format{
A \code{tibble} with 1781 rows and 11 variables:
\describe{
\item{symbol}{stock symbol at the New York Stock Exchange}
\item{period}{end of the considered time period}
\item{ebit}{earnings before interest and taxes in USD}
\item{ebt}{earnings before taxes in USD}
\item{rd}{investment in research and development in USD}
\item{ni}{net income in USD}
\item{inventory}{value of inventory in USD}
\item{cr}{cash ratio (cash reserves divided by liabilities) in percent}
\item{liabilities}{liabilities in USD}
\item{eps}{earnings per share (net income divided by shares) in USD}
\item{roe}{return on equity (net income divided by equity) in percent}
}
}
\source{
\url{https://www.kaggle.com/datasets/dgawlik/nyse/} (modified)
}
\usage{
data("metrics")
}
\description{
Fundamental stock metrics extracted from the annual SEC 10K fillings
(2012-2016). Dataset contains missing values.
}
\keyword{datasets}
