% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practicals.R
\name{practicals}
\alias{practicals}
\title{Start the practicals for an R course}
\usage{
practicals(id = getOption("rcourse_id", default = select()))
}
\arguments{
\item{id}{The course id.}
}
\value{
No return value.
}
\description{
Start the practicals for an R course
}
