% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prices}
\alias{prices}
\title{Stock prices}
\format{
A \code{tibble} with 851264 rows and 4 variables:
\describe{
\item{date}{the date}
\item{symbol}{stock symbol at the New York Stock Exchange}
\item{close}{the closing price}
\item{volume}{the trading volume}
}
}
\source{
\url{https://www.kaggle.com/datasets/dgawlik/nyse/} (modified)
}
\usage{
data("prices")
}
\description{
A dataset containing closing prices and trading volumes at the New York Stock
Exchange from 2012 to 2016. Most of the data spans from 2010 to the end 2016,
for companies new at the stock market the date range is shorter. There have
been approximately 140 stock splits in that time, the data set has been
adjusted accordingly.
}
\keyword{datasets}
