% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{RenderedDoc}
\alias{RenderedDoc}
\title{RenderedDoc}
\description{
...
}
\section{Super class}{
\code{\link[rooc:Doc]{rooc::Doc}} -> \code{RenderedDoc}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RenderedDoc-new}{\code{RenderedDoc$new()}}
\item \href{#method-RenderedDoc-rmdFile}{\code{RenderedDoc$rmdFile()}}
\item \href{#method-RenderedDoc-naviIds}{\code{RenderedDoc$naviIds()}}
\item \href{#method-RenderedDoc-type}{\code{RenderedDoc$type()}}
\item \href{#method-RenderedDoc-sessionIdx}{\code{RenderedDoc$sessionIdx()}}
\item \href{#method-RenderedDoc-lectureIdx}{\code{RenderedDoc$lectureIdx()}}
\item \href{#method-RenderedDoc-setupFun}{\code{RenderedDoc$setupFun()}}
\item \href{#method-RenderedDoc-hideFoldedCode}{\code{RenderedDoc$hideFoldedCode()}}
\item \href{#method-RenderedDoc-clone}{\code{RenderedDoc$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rooc" data-topic="Doc" data-id="id"><a href='../../rooc/html/Doc.html#method-Doc-id'><code>rooc::Doc$id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="Doc" data-id="label"><a href='../../rooc/html/Doc.html#method-Doc-label'><code>rooc::Doc$label()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="Doc" data-id="url"><a href='../../rooc/html/Doc.html#method-Doc-url'><code>rooc::Doc$url()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderedDoc-new"></a>}}
\if{latex}{\out{\hypertarget{method-RenderedDoc-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderedDoc$new(
  rmdFile = NULL,
  naviIds = NULL,
  type = "?",
  sessionIdx = 0L,
  lectureIdx = 0L,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderedDoc-rmdFile"></a>}}
\if{latex}{\out{\hypertarget{method-RenderedDoc-rmdFile}{}}}
\subsection{Method \code{rmdFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderedDoc$rmdFile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderedDoc-naviIds"></a>}}
\if{latex}{\out{\hypertarget{method-RenderedDoc-naviIds}{}}}
\subsection{Method \code{naviIds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderedDoc$naviIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderedDoc-type"></a>}}
\if{latex}{\out{\hypertarget{method-RenderedDoc-type}{}}}
\subsection{Method \code{type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderedDoc$type(long = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderedDoc-sessionIdx"></a>}}
\if{latex}{\out{\hypertarget{method-RenderedDoc-sessionIdx}{}}}
\subsection{Method \code{sessionIdx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderedDoc$sessionIdx()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderedDoc-lectureIdx"></a>}}
\if{latex}{\out{\hypertarget{method-RenderedDoc-lectureIdx}{}}}
\subsection{Method \code{lectureIdx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderedDoc$lectureIdx()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderedDoc-setupFun"></a>}}
\if{latex}{\out{\hypertarget{method-RenderedDoc-setupFun}{}}}
\subsection{Method \code{setupFun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderedDoc$setupFun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderedDoc-hideFoldedCode"></a>}}
\if{latex}{\out{\hypertarget{method-RenderedDoc-hideFoldedCode}{}}}
\subsection{Method \code{hideFoldedCode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderedDoc$hideFoldedCode()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenderedDoc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RenderedDoc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenderedDoc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
