% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{Renderer}
\alias{Renderer}
\title{Renderer}
\description{
...
}
\section{Super class}{
\code{\link[rooc:BaseRenderer]{rooc::BaseRenderer}} -> \code{Renderer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{Renderer$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="clearDir">}\href{../../rooc/html/BaseRenderer.html#method-clearDir}{\code{rooc::BaseRenderer$clearDir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="colorHtml">}\href{../../rooc/html/BaseRenderer.html#method-colorHtml}{\code{rooc::BaseRenderer$colorHtml()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="copyDoc">}\href{../../rooc/html/BaseRenderer.html#method-copyDoc}{\code{rooc::BaseRenderer$copyDoc()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="docUrl">}\href{../../rooc/html/BaseRenderer.html#method-docUrl}{\code{rooc::BaseRenderer$docUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="emojiHtml">}\href{../../rooc/html/BaseRenderer.html#method-emojiHtml}{\code{rooc::BaseRenderer$emojiHtml()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="extRefHtml">}\href{../../rooc/html/BaseRenderer.html#method-extRefHtml}{\code{rooc::BaseRenderer$extRefHtml()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="id2url">}\href{../../rooc/html/BaseRenderer.html#method-id2url}{\code{rooc::BaseRenderer$id2url()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="initialize">}\href{../../rooc/html/BaseRenderer.html#method-initialize}{\code{rooc::BaseRenderer$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="intRefHtml">}\href{../../rooc/html/BaseRenderer.html#method-intRefHtml}{\code{rooc::BaseRenderer$intRefHtml()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeAll">}\href{../../rooc/html/BaseRenderer.html#method-makeAll}{\code{rooc::BaseRenderer$makeAll()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeDoc">}\href{../../rooc/html/BaseRenderer.html#method-makeDoc}{\code{rooc::BaseRenderer$makeDoc()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeLectures">}\href{../../rooc/html/BaseRenderer.html#method-makeLectures}{\code{rooc::BaseRenderer$makeLectures()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeMaterials">}\href{../../rooc/html/BaseRenderer.html#method-makeMaterials}{\code{rooc::BaseRenderer$makeMaterials()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeTasks">}\href{../../rooc/html/BaseRenderer.html#method-makeTasks}{\code{rooc::BaseRenderer$makeTasks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="makeToc">}\href{../../rooc/html/BaseRenderer.html#method-makeToc}{\code{rooc::BaseRenderer$makeToc()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="mapOutHtmlFile">}\href{../../rooc/html/BaseRenderer.html#method-mapOutHtmlFile}{\code{rooc::BaseRenderer$mapOutHtmlFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="mapOutMaterialFile">}\href{../../rooc/html/BaseRenderer.html#method-mapOutMaterialFile}{\code{rooc::BaseRenderer$mapOutMaterialFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="mapTmpRmdFile">}\href{../../rooc/html/BaseRenderer.html#method-mapTmpRmdFile}{\code{rooc::BaseRenderer$mapTmpRmdFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="outDir">}\href{../../rooc/html/BaseRenderer.html#method-outDir}{\code{rooc::BaseRenderer$outDir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="renderDoc">}\href{../../rooc/html/BaseRenderer.html#method-renderDoc}{\code{rooc::BaseRenderer$renderDoc()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rooc" data-topic="BaseRenderer" data-id="specialBlockHtml">}\href{../../rooc/html/BaseRenderer.html#method-specialBlockHtml}{\code{rooc::BaseRenderer$specialBlockHtml()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
