% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{slides}
\alias{slides}
\title{Open the slides of an R course}
\usage{
slides(id = getOption("rcourse_id", default = select()))
}
\arguments{
\item{id}{The course id.}
}
\value{
No return value.
}
\description{
Open the slides of an R course
}
