% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Stocks}
\format{
A \code{tibble} with 505 rows and 6 variables:
\describe{
\item{symbol}{stock symbol at the New York Stock Exchange}
\item{name}{the name of the security}
\item{sector}{the GICS sector}
\item{industry}{the GICS sub-industry}
\item{hq}{the state/country of the headquarter}
\item{added}{the date the stock was first added to the S&P 500}
}
}
\source{
\url{https://www.kaggle.com/datasets/dgawlik/nyse/} (modified)
}
\usage{
data("stocks")
}
\description{
A dataset containing general descriptions of stocks in the American S&P 500
index from 2012 to 2016.
}
\keyword{datasets}
