% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{TheCourse}
\alias{TheCourse}
\title{TheCourse}
\description{
...
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TheCourse-new}{\code{TheCourse$new()}}
\item \href{#method-TheCourse-id}{\code{TheCourse$id()}}
\item \href{#method-TheCourse-dir}{\code{TheCourse$dir()}}
\item \href{#method-TheCourse-label}{\code{TheCourse$label()}}
\item \href{#method-TheCourse-sessionIds}{\code{TheCourse$sessionIds()}}
\item \href{#method-TheCourse-lectureIds}{\code{TheCourse$lectureIds()}}
\item \href{#method-TheCourse-lectureWithTasksIds}{\code{TheCourse$lectureWithTasksIds()}}
\item \href{#method-TheCourse-materialIds}{\code{TheCourse$materialIds()}}
\item \href{#method-TheCourse-id2label}{\code{TheCourse$id2label()}}
\item \href{#method-TheCourse-add}{\code{TheCourse$add()}}
\item \href{#method-TheCourse-materialsTibble}{\code{TheCourse$materialsTibble()}}
\item \href{#method-TheCourse-lecturesTibble}{\code{TheCourse$lecturesTibble()}}
\item \href{#method-TheCourse-lectureDoc}{\code{TheCourse$lectureDoc()}}
\item \href{#method-TheCourse-taskDoc}{\code{TheCourse$taskDoc()}}
\item \href{#method-TheCourse-tocDoc}{\code{TheCourse$tocDoc()}}
\item \href{#method-TheCourse-materialsDoc}{\code{TheCourse$materialsDoc()}}
\item \href{#method-TheCourse-materialDoc}{\code{TheCourse$materialDoc()}}
\item \href{#method-TheCourse-clone}{\code{TheCourse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-new"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$new(id, dir, label = id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-id"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-id}{}}}
\subsection{Method \code{id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-dir"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-dir}{}}}
\subsection{Method \code{dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$dir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-label"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-label}{}}}
\subsection{Method \code{label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$label()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-sessionIds"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-sessionIds}{}}}
\subsection{Method \code{sessionIds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$sessionIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-lectureIds"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-lectureIds}{}}}
\subsection{Method \code{lectureIds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$lectureIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-lectureWithTasksIds"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-lectureWithTasksIds}{}}}
\subsection{Method \code{lectureWithTasksIds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$lectureWithTasksIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-materialIds"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-materialIds}{}}}
\subsection{Method \code{materialIds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$materialIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-id2label"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-id2label}{}}}
\subsection{Method \code{id2label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$id2label(docId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-add"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$add(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-materialsTibble"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-materialsTibble}{}}}
\subsection{Method \code{materialsTibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$materialsTibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-lecturesTibble"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-lecturesTibble}{}}}
\subsection{Method \code{lecturesTibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$lecturesTibble(withBreaks = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-lectureDoc"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-lectureDoc}{}}}
\subsection{Method \code{lectureDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$lectureDoc(lectureId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-taskDoc"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-taskDoc}{}}}
\subsection{Method \code{taskDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$taskDoc(lectureId, enableCode)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-tocDoc"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-tocDoc}{}}}
\subsection{Method \code{tocDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$tocDoc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-materialsDoc"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-materialsDoc}{}}}
\subsection{Method \code{materialsDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$materialsDoc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-materialDoc"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-materialDoc}{}}}
\subsection{Method \code{materialDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$materialDoc(materialId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TheCourse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TheCourse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
