% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{BaseRenderer}
\alias{BaseRenderer}
\title{BaseRenderer}
\description{
...
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BaseRenderer$new()}}
\item \href{#method-outDir}{\code{BaseRenderer$outDir()}}
\item \href{#method-colorHtml}{\code{BaseRenderer$colorHtml()}}
\item \href{#method-emojiHtml}{\code{BaseRenderer$emojiHtml()}}
\item \href{#method-specialBlockHtml}{\code{BaseRenderer$specialBlockHtml()}}
\item \href{#method-extRefHtml}{\code{BaseRenderer$extRefHtml()}}
\item \href{#method-intRefHtml}{\code{BaseRenderer$intRefHtml()}}
\item \href{#method-id2url}{\code{BaseRenderer$id2url()}}
\item \href{#method-makeAll}{\code{BaseRenderer$makeAll()}}
\item \href{#method-makeToc}{\code{BaseRenderer$makeToc()}}
\item \href{#method-makeMaterials}{\code{BaseRenderer$makeMaterials()}}
\item \href{#method-makeLectures}{\code{BaseRenderer$makeLectures()}}
\item \href{#method-makeTasks}{\code{BaseRenderer$makeTasks()}}
\item \href{#method-docUrl}{\code{BaseRenderer$docUrl()}}
\item \href{#method-mapTmpRmdFile}{\code{BaseRenderer$mapTmpRmdFile()}}
\item \href{#method-mapOutHtmlFile}{\code{BaseRenderer$mapOutHtmlFile()}}
\item \href{#method-mapOutMaterialFile}{\code{BaseRenderer$mapOutMaterialFile()}}
\item \href{#method-clearDir}{\code{BaseRenderer$clearDir()}}
\item \href{#method-makeDoc}{\code{BaseRenderer$makeDoc()}}
\item \href{#method-copyDoc}{\code{BaseRenderer$copyDoc()}}
\item \href{#method-renderDoc}{\code{BaseRenderer$renderDoc()}}
\item \href{#method-clone}{\code{BaseRenderer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$new(outDir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-outDir"></a>}}
\if{latex}{\out{\hypertarget{method-outDir}{}}}
\subsection{Method \code{outDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$outDir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-colorHtml"></a>}}
\if{latex}{\out{\hypertarget{method-colorHtml}{}}}
\subsection{Method \code{colorHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$colorHtml(..., color)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-emojiHtml"></a>}}
\if{latex}{\out{\hypertarget{method-emojiHtml}{}}}
\subsection{Method \code{emojiHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$emojiHtml(id, text = id, color = "red")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-specialBlockHtml"></a>}}
\if{latex}{\out{\hypertarget{method-specialBlockHtml}{}}}
\subsection{Method \code{specialBlockHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$specialBlockHtml(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extRefHtml"></a>}}
\if{latex}{\out{\hypertarget{method-extRefHtml}{}}}
\subsection{Method \code{extRefHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$extRefHtml(..., url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-intRefHtml"></a>}}
\if{latex}{\out{\hypertarget{method-intRefHtml}{}}}
\subsection{Method \code{intRefHtml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$intRefHtml(..., url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-id2url"></a>}}
\if{latex}{\out{\hypertarget{method-id2url}{}}}
\subsection{Method \code{id2url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$id2url(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeAll"></a>}}
\if{latex}{\out{\hypertarget{method-makeAll}{}}}
\subsection{Method \code{makeAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeAll(course, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeToc"></a>}}
\if{latex}{\out{\hypertarget{method-makeToc}{}}}
\subsection{Method \code{makeToc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeToc(course, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeMaterials"></a>}}
\if{latex}{\out{\hypertarget{method-makeMaterials}{}}}
\subsection{Method \code{makeMaterials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeMaterials(course, materialIds = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeLectures"></a>}}
\if{latex}{\out{\hypertarget{method-makeLectures}{}}}
\subsection{Method \code{makeLectures()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeLectures(course, lectureIds = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeTasks"></a>}}
\if{latex}{\out{\hypertarget{method-makeTasks}{}}}
\subsection{Method \code{makeTasks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeTasks(course, enableCode, lectureIds = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-docUrl"></a>}}
\if{latex}{\out{\hypertarget{method-docUrl}{}}}
\subsection{Method \code{docUrl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$docUrl(doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapTmpRmdFile"></a>}}
\if{latex}{\out{\hypertarget{method-mapTmpRmdFile}{}}}
\subsection{Method \code{mapTmpRmdFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$mapTmpRmdFile(doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapOutHtmlFile"></a>}}
\if{latex}{\out{\hypertarget{method-mapOutHtmlFile}{}}}
\subsection{Method \code{mapOutHtmlFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$mapOutHtmlFile(doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapOutMaterialFile"></a>}}
\if{latex}{\out{\hypertarget{method-mapOutMaterialFile}{}}}
\subsection{Method \code{mapOutMaterialFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$mapOutMaterialFile(doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clearDir"></a>}}
\if{latex}{\out{\hypertarget{method-clearDir}{}}}
\subsection{Method \code{clearDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$clearDir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeDoc"></a>}}
\if{latex}{\out{\hypertarget{method-makeDoc}{}}}
\subsection{Method \code{makeDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$makeDoc(course, doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copyDoc"></a>}}
\if{latex}{\out{\hypertarget{method-copyDoc}{}}}
\subsection{Method \code{copyDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$copyDoc(course, doc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renderDoc"></a>}}
\if{latex}{\out{\hypertarget{method-renderDoc}{}}}
\subsection{Method \code{renderDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$renderDoc(course, doc, quiet = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseRenderer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
