% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{CopiedDoc}
\alias{CopiedDoc}
\title{CopiedDoc}
\description{
...
}
\section{Super class}{
\code{\link[rooc:Doc]{rooc::Doc}} -> \code{CopiedDoc}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CopiedDoc-new}{\code{CopiedDoc$new()}}
\item \href{#method-CopiedDoc-path}{\code{CopiedDoc$path()}}
\item \href{#method-CopiedDoc-outPath}{\code{CopiedDoc$outPath()}}
\item \href{#method-CopiedDoc-url}{\code{CopiedDoc$url()}}
\item \href{#method-CopiedDoc-clone}{\code{CopiedDoc$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rooc" data-topic="Doc" data-id="id"><a href='../../rooc/html/Doc.html#method-Doc-id'><code>rooc::Doc$id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rooc" data-topic="Doc" data-id="label"><a href='../../rooc/html/Doc.html#method-Doc-label'><code>rooc::Doc$label()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CopiedDoc-new"></a>}}
\if{latex}{\out{\hypertarget{method-CopiedDoc-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopiedDoc$new(path, outPath = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CopiedDoc-path"></a>}}
\if{latex}{\out{\hypertarget{method-CopiedDoc-path}{}}}
\subsection{Method \code{path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopiedDoc$path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CopiedDoc-outPath"></a>}}
\if{latex}{\out{\hypertarget{method-CopiedDoc-outPath}{}}}
\subsection{Method \code{outPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopiedDoc$outPath()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CopiedDoc-url"></a>}}
\if{latex}{\out{\hypertarget{method-CopiedDoc-url}{}}}
\subsection{Method \code{url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopiedDoc$url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CopiedDoc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CopiedDoc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CopiedDoc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
