% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{Lecture}
\alias{Lecture}
\title{Lecture}
\description{
...
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lecture-new}{\code{Lecture$new()}}
\item \href{#method-Lecture-id}{\code{Lecture$id()}}
\item \href{#method-Lecture-label}{\code{Lecture$label()}}
\item \href{#method-Lecture-rmdFile}{\code{Lecture$rmdFile()}}
\item \href{#method-Lecture-hasTasks}{\code{Lecture$hasTasks()}}
\item \href{#method-Lecture-min}{\code{Lecture$min()}}
\item \href{#method-Lecture-asTibble}{\code{Lecture$asTibble()}}
\item \href{#method-Lecture-clone}{\code{Lecture$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lecture-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lecture-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$new(id, label = NULL, rmdFile = NULL, hasTasks = TRUE, min = NA, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lecture-id"></a>}}
\if{latex}{\out{\hypertarget{method-Lecture-id}{}}}
\subsection{Method \code{id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lecture-label"></a>}}
\if{latex}{\out{\hypertarget{method-Lecture-label}{}}}
\subsection{Method \code{label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$label()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lecture-rmdFile"></a>}}
\if{latex}{\out{\hypertarget{method-Lecture-rmdFile}{}}}
\subsection{Method \code{rmdFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$rmdFile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lecture-hasTasks"></a>}}
\if{latex}{\out{\hypertarget{method-Lecture-hasTasks}{}}}
\subsection{Method \code{hasTasks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$hasTasks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lecture-min"></a>}}
\if{latex}{\out{\hypertarget{method-Lecture-min}{}}}
\subsection{Method \code{min()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$min()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lecture-asTibble"></a>}}
\if{latex}{\out{\hypertarget{method-Lecture-asTibble}{}}}
\subsection{Method \code{asTibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$asTibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lecture-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lecture-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lecture$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
