% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{Session}
\alias{Session}
\title{Session}
\description{
...
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Session-new}{\code{Session$new()}}
\item \href{#method-Session-id}{\code{Session$id()}}
\item \href{#method-Session-lectureIds}{\code{Session$lectureIds()}}
\item \href{#method-Session-add}{\code{Session$add()}}
\item \href{#method-Session-lecturesTibble}{\code{Session$lecturesTibble()}}
\item \href{#method-Session-clone}{\code{Session$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-new"></a>}}
\if{latex}{\out{\hypertarget{method-Session-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$new(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-id"></a>}}
\if{latex}{\out{\hypertarget{method-Session-id}{}}}
\subsection{Method \code{id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-lectureIds"></a>}}
\if{latex}{\out{\hypertarget{method-Session-lectureIds}{}}}
\subsection{Method \code{lectureIds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$lectureIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-add"></a>}}
\if{latex}{\out{\hypertarget{method-Session-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$add(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-lecturesTibble"></a>}}
\if{latex}{\out{\hypertarget{method-Session-lecturesTibble}{}}}
\subsection{Method \code{lecturesTibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$lecturesTibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Session-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
