% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\name{TheCourse}
\alias{TheCourse}
\title{TheCourse}
\description{
...
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TheCourse$new()}}
\item \href{#method-id}{\code{TheCourse$id()}}
\item \href{#method-dir}{\code{TheCourse$dir()}}
\item \href{#method-label}{\code{TheCourse$label()}}
\item \href{#method-sessionIds}{\code{TheCourse$sessionIds()}}
\item \href{#method-lectureIds}{\code{TheCourse$lectureIds()}}
\item \href{#method-lectureWithTasksIds}{\code{TheCourse$lectureWithTasksIds()}}
\item \href{#method-materialIds}{\code{TheCourse$materialIds()}}
\item \href{#method-id2label}{\code{TheCourse$id2label()}}
\item \href{#method-add}{\code{TheCourse$add()}}
\item \href{#method-materialsTibble}{\code{TheCourse$materialsTibble()}}
\item \href{#method-lecturesTibble}{\code{TheCourse$lecturesTibble()}}
\item \href{#method-lectureDoc}{\code{TheCourse$lectureDoc()}}
\item \href{#method-taskDoc}{\code{TheCourse$taskDoc()}}
\item \href{#method-tocDoc}{\code{TheCourse$tocDoc()}}
\item \href{#method-materialsDoc}{\code{TheCourse$materialsDoc()}}
\item \href{#method-materialDoc}{\code{TheCourse$materialDoc()}}
\item \href{#method-clone}{\code{TheCourse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$new(id, dir, label = id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-id"></a>}}
\if{latex}{\out{\hypertarget{method-id}{}}}
\subsection{Method \code{id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dir"></a>}}
\if{latex}{\out{\hypertarget{method-dir}{}}}
\subsection{Method \code{dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$dir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label"></a>}}
\if{latex}{\out{\hypertarget{method-label}{}}}
\subsection{Method \code{label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$label()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sessionIds"></a>}}
\if{latex}{\out{\hypertarget{method-sessionIds}{}}}
\subsection{Method \code{sessionIds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$sessionIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lectureIds"></a>}}
\if{latex}{\out{\hypertarget{method-lectureIds}{}}}
\subsection{Method \code{lectureIds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$lectureIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lectureWithTasksIds"></a>}}
\if{latex}{\out{\hypertarget{method-lectureWithTasksIds}{}}}
\subsection{Method \code{lectureWithTasksIds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$lectureWithTasksIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-materialIds"></a>}}
\if{latex}{\out{\hypertarget{method-materialIds}{}}}
\subsection{Method \code{materialIds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$materialIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-id2label"></a>}}
\if{latex}{\out{\hypertarget{method-id2label}{}}}
\subsection{Method \code{id2label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$id2label(docId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$add(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-materialsTibble"></a>}}
\if{latex}{\out{\hypertarget{method-materialsTibble}{}}}
\subsection{Method \code{materialsTibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$materialsTibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lecturesTibble"></a>}}
\if{latex}{\out{\hypertarget{method-lecturesTibble}{}}}
\subsection{Method \code{lecturesTibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$lecturesTibble(withBreaks = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lectureDoc"></a>}}
\if{latex}{\out{\hypertarget{method-lectureDoc}{}}}
\subsection{Method \code{lectureDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$lectureDoc(lectureId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-taskDoc"></a>}}
\if{latex}{\out{\hypertarget{method-taskDoc}{}}}
\subsection{Method \code{taskDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$taskDoc(lectureId, enableCode)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tocDoc"></a>}}
\if{latex}{\out{\hypertarget{method-tocDoc}{}}}
\subsection{Method \code{tocDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$tocDoc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-materialsDoc"></a>}}
\if{latex}{\out{\hypertarget{method-materialsDoc}{}}}
\subsection{Method \code{materialsDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$materialsDoc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-materialDoc"></a>}}
\if{latex}{\out{\hypertarget{method-materialDoc}{}}}
\subsection{Method \code{materialDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$materialDoc(materialId)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TheCourse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
